apply mild opacity to podcast player when in dark mode
