+++ packages/lesswrong/lib/collections/users/fragments.ts
@@ -153,6 +153,7 @@ registerFragment(`
     shortformFeedId
     viewUnreviewedComments
     recommendationSettings
+    theme
 
     bookmarkedPostsMetadata
